pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "Cython>=3.0"]
build-backend = "setuptools.build_meta"

[project]
name = "1c-processor-generator"
version = "2.62.2"
description = "Automatic generation of external data processors for 1C:Enterprise 8.3"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT (Open Source) + Proprietary (PRO Module)"}
authors = [
    {name = "ITDEO", email = "itdeo.tech@gmail.com"}
]
keywords = ["1c", "enterprise", "generator", "bsl", "yaml", "automation", "epf"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Code Generators",
    "License :: OSI Approved :: MIT License",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    # Core
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
    # Sync tool (v2.25.0+)
    "ruamel.yaml>=0.17.0",
    "lxml>=4.9.0",
    # PRO module (v2.51.0+)
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
# SVG to PNG conversion (requires cairo system library on Windows)
svg = [
    "cairosvg>=2.7.0",
    "Pillow>=10.0.0",
]
# Excel to MXL conversion (v2.58.0+)
excel = [
    "openpyxl>=3.1.0",
]
# All optional features
all = [
    "cairosvg>=2.7.0",
    "Pillow>=10.0.0",
    "openpyxl>=3.1.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
# Use wildcard because package name starts with digit (invalid Python identifier)
"*" = [
    "*.pyd",
    "*.json",
    "templates/*.j2",
    "templates/**/*.j2",
    "_protected_bins/**/*.pyd",
    "pro/*.pyd",
    "pro/LICENSE",
    "pro/_bins/**/*.pyd",
]

[project.urls]
Homepage = "https://itdeo.tech/1c-processor-generator"
Documentation = "https://itdeo.tech/1c-processor-generator"
Repository = "https://github.com/SERGIUSH-UA/1c-processor-generator-pro"
"Bug Tracker" = "https://github.com/SERGIUSH-UA/1c-processor-generator-pro/issues"
"Buy PRO" = "https://itdeo.tech/1c-processor-generator/#pricing"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=1c_processor_generator",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["1c_processor_generator"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
